+++ test/unit/Nautilus.test.ts
@@ -1,5 +1,5 @@
 import { Nautilus } from '../../src'
-import { getSigner } from '../fixtures/Web3'
+import { getSigner } from '../fixtures/Ethers'
 import assert from 'assert'
 import { LoggerInstance, LogLevel } from '@oceanprotocol/lib'
